requirements.txt
Django==6.0.3
djangorestframework==3.16.1
django-allauth==65.15.0
django-cors-headers==4.9.0
django-environ==0.13.0
django-storages==1.14.6
channels==4.3.2
channels_redis==4.3.0
daphne==4.2.1
celery==5.6.2
redis==7.3.0
psycopg2-binary==2.9.11
boto3==1.42.68
gunicorn==23.0.0
